renaming loaders and trpc file location
