+++ src/routes/board/layout.tsx
@@ -4,7 +4,7 @@ import { ProtectedHeader } from "~/modules/layout/ProtectedHeader/ProtectedHeade
 import { getUserFromEvent } from "~/server/loaders";
 import { paths } from "~/utils/paths";
 
-export const useProtectedUserLoader = loader$(async (event) => {
+export const useProtectedRoute = loader$(async (event) => {
   const user = await getUserFromEvent(event);
 
   if (!user) {
@@ -13,7 +13,7 @@ export const useProtectedUserLoader = loader$(async (event) => {
 });
 
 export default component$(() => {
-  useProtectedUserLoader();
+  useProtectedRoute();
   return (
     <>
       <ProtectedHeader />
